Standardize lcgmake binary name on "lcgmake".
Some scripts (e.g. Wiking's Makefile on "make doc") expect lcgmake script to
have exactly that name and be available in the PATH. This commit removes the
need to create symlinks with the expected name manually for these scripts to
work without the need to install lcg system-wide (either with "make install"
or "make install-links").