Support for embedding YouTube and Vimeo videos
Both links and URIs to YouTube and Vimeo pages
containing videos are recognized by the parser and
in HTML rendered as embedded players.

The HTML generator needed to be extended to
handle the OBJECT and PARAM tags as described
in HTML4.

I've considered including the objects using swfobject.js
but decided against it as it would require additional
javascript methods and a rework of embed_swf_object()
which currently only supports flash objects
from resources, not external URIs. If Flash is not available
on client side, the browser is responsible for handling
the situation and generally it will warn the user and ask
him whether he wants to install Flash.