Handle sections 'Container' to get rid of 'SectionContainer'.
The class 'SectionContainer' existed for historical reasons and it was totally
unpractical to decide which kind of contaier is needed based on its content.
This commit removes the need for 'SectionContainer' alltogether and also drops
the support for automatic 'TableOfContents' creation supported by this
container as this feature was actually never used anywhere else than in
Eurochance courses.

The primary motivation for this change was the traceback in 'SectionContainer'
processing after recent changes in the export mechanism.  This traceback was
hard to fix as the automatic 'TableOfContents' creation was an inherently wrong
idea.

The identifier 'SectionContainer' was kept for backwards compatibility, but it
works exactly as ordinary 'Container' now.  Eurochance courses will need to add
'TableOfContents' explicitly where needed.