Determine export element type automatically.
This commit slightly modifies the way how export of generic content elements is
implemented.  The element type is now recognized by the exporter automatically,
so the element doesn't need to pass its type to `Exporter.export_element()'
explicitly.  The type is simply determined as the element's class itself or the
first class from the element's base classes, which is supported by the
exporter.  This eliminates the need to override
`Content._export_element_type()' in classes derived from generic content
elements defined by LCG.