Added support for datetime conversions during localization
The time zone can now be passed to the export context and LocalizableDateTime
values will be automatically converted to given time zone when they are created
with the 'utc' flag set to true (false by default).

As the time zone is independent of the language, the 'Translator' class had to be
replaced with a more generic 'Localizer' class which now does all
localization (translations + locale dependent formatting + time zone
conversions).

The changes should be backwards compatible as there are aliases for
'Exporter.translator()' (now deprecated) and 'Localizer.translate()'.
'Exporter.translator()'will actually return a 'Localizer' instance, which is,
however API compatible with the former 'Translator', except for methods used by
LCG internally.