Fix for iBooks behavior with XML and spaces
iBooks seems to parse all XML as if xml:space="preserve". So we do not
include any space that is no part of content in the XML documents,
which makes the XML hard to read as text by humans. Maybe I will try
to set explicitely xml:space="default" to see whether that would work,
but I doubt it will.