Implement support for interactive exercises
The usege is currently limited to content edited through an HTML editor or to
manual construction of Exercise instances.  The functionality is taken from the
Eurochance extension module with some modifications and larger code
refactorings.  The JavaScript Code was rewitten to make use of Prototype.js and
the existing LCG JavaScript classes.