Fix adding pybabel to Travis
Python dependencies in .travis.yml should be installed by pip, not
through the distribution. See
http://docs.travis-ci.com/user/languages/python/#Travis-CI-Uses-Isolated-virtualenvs
for explanation.

This currently works anyway only because we use the same python version
as is default for the Ubuntu release on which Travis CI runs. Should we
use more Python versions for testing, installing through apt would
break.