Test Parser by comparing HTML exports
Comparing the HTML exports seems a more precise comparison than a bunch
of assertions.  Plus the feedback on failure gives a much better idea
about the problem.

Also this is one of the first tests relying on pytest, So we are using a
simple assert statement instead of self.assertEqual as it gives a better
feedback.