Allow passing id to Container
The argument 'id' in Container constructor was kept for backwards
compatibility as an alias of 'name'.  As the old variant is not used
anymore, we can now finally change its meaning to be consistent with
this argument of Section.  Thus all containers (including derived
classes such as Emphasized) now can have both 'id' and 'name'.