Distinguish collapsible pane/section
Create a new collapsible widget which is not a section.  The original
'CollapsiblePane' class is now renamed to 'CollapsibleSection'.  The new
class uses the old name 'CollapsiblePane'.  The two widgets differ in
their HTML representation.  'CollapsibleSection' uses HTML H tag for the
heading, 'CollapsiblePane' used a simple div with a link.