Handle TOC back references in export context
The initial motivation for this change was getting rid of the configuration
variable allow_backref.  When at it, handling this was moved to a much more
appropriate place because having it in content classes was a big old hack.

Support for back references in PDF export was dropped because this feature is
just not very useful there.