Always produce invariant PDFs
Without this, Reportlab adds a unique signature at the end of each PDF.  This
makes the PDF hard to compare (for example using a file checksum).  As we don't
have any use for the unique signatures, we prefer to always have invariant
output.  If we ever realize we need unique output, we may add an option
somewhere.