Rename the package to 'lcg-framework'
PyPi refuses the package name 'lcg' even if the name is not taken by another
project.  The page https://pypi.org/help/#project-name explains that short
names are discouraged.  It seems that exceptions are rarely accepted so we will
probably need to live with package name not matching the library name.  The
library is still imported like "import lcg", but installation will be "pip
install lcg-framework".